:dizzy: factor out digest.js
