+++ bin/start.js
@@ -1,3 +1,3 @@
 #!/usr/bin/env node
 
-require('../server')();
+require('../lib/server')();
