+++ lib/db.js
@@ -1,4 +1,4 @@
-const config = require('./config');
+const config = require('../config');
 const pg = require('pg');
 const db = new pg.Client(config.db_dsn);
 db.connect();
