+++ lib/server.js
@@ -3,8 +3,8 @@ module.exports = () => {
 	let server;
 
 	const fs = require('fs');
-	const config = require('./config');
-	const app = require('./app');
+	const config = require('../config');
+	const app = require('../app');
 
 	if ('ssl' in config) {
 		// Setup HTTPS server
