+++ test/server.js
@@ -1,18 +1,18 @@
 var request = require('supertest');
-describe('loading express', function () {
+describe('loading express', () => {
 	var server;
-	beforeEach(function () {
-		server = require('../server')();
+	beforeEach(() => {
+		server = require('../lib/server')();
 	});
-	afterEach(function () {
+	afterEach(() => {
 		server.close();
 	});
-	it('responds to /', function test_home(done) {
+	it('responds to /', (done) => {
 		request(server)
 		.get('/')
 		.expect(200, done);
 	});
-	it('404 everything else', function test_404(done) {
+	it('404 everything else', (done) => {
 		request(server)
 		.get('/foo/bar')
 		.expect(404, done);
